+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update entrypoint to skip custom PAM module setup for consumer installs (#52)
 - Update entrypoint to clarify iRODS server wait messages (#53)
 - Add `IRODS_SERVICE_ACCOUNT_UID` and `IRODS_SERVICE_ACCOUNT_GID` env vars (#50)
+- Move service user and resource directory setup in entrypoint (#54)
 
 
 ## v4.3.3-1 (2024-12-19)